HW13/requirements.txt
streamlit
tensorflow
gdown
pillow
numpy
requests
matplotlib
scikit-learn
scipy
pandas
seaborn
scikit-image
opencv-python
tqdm
plotly
pydub
moviepy